ai-service/app/services/fraud_detector.py
import cv2
import numpy as np
import requests
from PIL import Image
from io import BytesIO
import logging

logger = logging.getLogger(__name__)

class FraudDetector:
    def __init__(self):
        self.min_resolution = (200, 200)
        self.max_compression_artifacts = 50
    
    async def detect(self, image_url: str) -> float:
        try:
            image = await self._download_image(image_url)
            
            scores = []
            scores.append(self._check_resolution(image))
            scores.append(self._check_metadata(image))
            scores.append(self._check_compression(image))
            scores.append(self._check_duplicates(image))
            
            fraud_score = sum(scores) / len(scores)
            return fraud_score
        except Exception as e:
            logger.error(f"Fraud detection error: {str(e)}")
            return 0.5
    
    async def _download_image(self, url: str):
        response = requests.get(url, timeout=10)
        image = Image.open(BytesIO(response.content))
        return cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
    
    def _check_resolution(self, image) -> float:
        h, w = image.shape[:2]
        if w < self.min_resolution[0] or h < self.min_resolution[1]:
            return 0.8
        return 0.0
    
    def _check_metadata(self, image) -> float:
        return 0.0
    
    def _check_compression(self, image) -> float:
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        laplacian_var = cv2.Laplacian(gray, cv2.CV_64F).var()
        
        if laplacian_var < 100:
            return 0.6
        return 0.0
    
    def _check_duplicates(self, image) -> float:
        return 0.0
