ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.2
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.227
